Enhance cleanDir function to prevent removal of TEXCACHEROOT, improve error logging in transferTexFileToHost, and reset caches in flushcache command

+++ dvipdfm.wasm/pre.js
@@ -84,7 +84,7 @@ function cleanDir(dir) {
         }
     }
 
-    if (dir !== WORKROOT) {
+    if (dir !== WORKROOT && dir !== TEXCACHEROOT) {
         try {
             FS.rmdir(dir);
         } catch (err) {
@@ -223,7 +223,7 @@ function transferTexFileToHost(filename) {
             'content': content
         }, [content.buffer]);
     } catch (err) {
-        console.error("Unable to fetch mem file", err);
+        console.error("Unable to fetch mem file " + filename, err);
         self.postMessage({
             'result': 'failed',
             'cmd': 'fetchfile'
@@ -299,6 +299,7 @@ self['onmessage'] = function (ev) {
     } else if (cmd === "flushcache") {
         cleanDir(TEXCACHEROOT);
         cleanDir(WORKROOT);
+        texlive200_cache = {};
         self.postMessage({
             'result': 'ok',
             'cmd': 'flushcache'
@@ -318,8 +319,6 @@ self['onmessage'] = function (ev) {
     } else if (cmd === "writecache") {
         texlive404_cache = data['texlive404_cache'];
         texlive200_cache = data['texlive200_cache'];
-        font404_cache = data['font404_cache'];
-        font200_cache = data['font200_cache'];
         self.postMessage({
             result: 'ok',
             cmd: 'writecache'
